# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: M.P. Warehousing & Logistics Corporation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/commodities:
    get:
      operationId: listCommodities
      tags: [commodities]
      summary: List commodities
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of commodities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commodity"
    post:
      operationId: createCommodity
      tags: [commodities]
      summary: Create commodity (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommodityRequest"
      responses:
        "201":
          description: Commodity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commodity"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/commodities/{id}:
    get:
      operationId: getCommodity
      tags: [commodities]
      summary: Get commodity by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Commodity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commodity"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCommodity
      tags: [commodities]
      summary: Update commodity (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommodityRequest"
      responses:
        "200":
          description: Updated commodity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commodity"
...
  /v1/bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills with optional filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: branch_name
          schema:
            type: string
        - in: query
          name: month_year
          schema:
            type: string
        - in: query
          name: commodity_id
          schema:
            type: integer
        - in: query
          name: created_by
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillListResponse"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill (operator only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillRequest"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get bill by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/bills/{id}/request-edit:
    post:
      operationId: requestBillEdit
      tags: [bills]
      summary: Request edit on a bill (operator only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestEditRequest"
      responses:
        "201":
          description: Edit request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillVersion"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/bills/{id}/request-delete:
    post:
      operationId: requestBillDelete
      tags: [bills]
      summary: Request delete on a bill (operator only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Delete request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillVersion"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/bills/{id}/versions:
    get:
      operationId: getBillVersions
      tags: [bills]
      summary: Get versions/edit requests for a bill
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of bill versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillVersion"
...
  /v1/bills/export:
    get:
      operationId: exportBills
      tags: [bills]
      summary: Export bills to CSV (admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: branch_name
          schema:
            type: string
        - in: query
          name: month_year
          schema:
            type: string
      responses:
        "200":
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operator]
        branch_name:
          type: ["string", "null"]
        district_name:
          type: ["string", "null"]
        mobile_number:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - created_at
...
    Commodity:
      type: object
      properties:
        id:
          type: integer
        crop_name:
          type: string
        crop_year:
          type: string
        per_bag_per_month:
          type: number
      required:
        - id
        - crop_name
        - crop_year
        - per_bag_per_month
    CreateCommodityRequest:
      type: object
      properties:
        crop_name:
          type: string
        crop_year:
          type: string
        per_bag_per_month:
          type: number
      required:
        - crop_name
        - crop_year
        - per_bag_per_month
    UpdateCommodityRequest:
      type: object
      properties:
        crop_name:
          type: string
        crop_year:
          type: string
        per_bag_per_month:
          type: number
    Depositor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gst_no:
          type: ["string", "null"]
        total_gst:
          type: ["number", "null"]
      required:
        - id
        - name
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        serial_no:
          type: integer
        created_by:
          type: integer
        district:
          type: ["string", "null"]
        branch_name:
          type: ["string", "null"]
        godown_name:
          type: ["string", "null"]
        bill_no:
          type: ["string", "null"]
        commodity_id:
          type: integer
        crop_year:
          type: ["string", "null"]
        financial_year:
          type: ["string", "null"]
        month_year:
          type: ["string", "null"]
        rate_per_bag:
          type: ["number", "null"]
        opening_balance:
          type: ["integer", "null"]
        closing_balance:
          type: ["integer", "null"]
        received_bags:
          type: ["integer", "null"]
        issue_bags:
          type: ["integer", "null"]
        reserve_bags:
          type: ["integer", "null"]
        chargeable_bags:
          type: ["integer", "null"]
        total_charge:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        is_locked:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - serial_no
        - created_by
        - commodity_id
        - status
        - is_locked
        - created_at
    BillDetail:
      type: object
      properties:
        id:
          type: integer
        serial_no:
          type: integer
        created_by:
          type: integer
        district:
          type: ["string", "null"]
        branch_name:
          type: ["string", "null"]
        godown_name:
          type: ["string", "null"]
        bill_no:
          type: ["string", "null"]
        commodity_id:
          type: integer
        crop_year:
          type: ["string", "null"]
        financial_year:
          type: ["string", "null"]
        month_year:
          type: ["string", "null"]
        rate_per_bag:
          type: ["number", "null"]
        opening_balance:
          type: ["integer", "null"]
        closing_balance:
          type: ["integer", "null"]
        received_bags:
          type: ["integer", "null"]
        issue_bags:
          type: ["integer", "null"]
        reserve_bags:
          type: ["integer", "null"]
        chargeable_bags:
          type: ["integer", "null"]
        total_charge:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        is_locked:
          type: boolean
        created_at:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/User"
        commodity:
          $ref: "#/components/schemas/Commodity"
        approval:
          oneOf:
            - $ref: "#/components/schemas/BillApproval"
            - type: "null"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/BillVersion"
      required:
        - id
        - serial_no
        - created_by
        - commodity_id
        - status
        - is_locked
        - created_at
        - creator
        - commodity
        - versions
    BillListResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/BillDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
      required:
        - bills
        - total
        - page
        - limit
        - total_pages
    CreateBillRequest:
      type: object
      properties:
        district:
          type: string
        branch_name:
          type: string
        godown_name:
          type: string
        bill_no:
          type: string
        commodity_id:
          type: integer
        crop_year:
          type: string
        financial_year:
          type: string
        month_year:
          type: string
        rate_per_bag:
          type: number
        opening_balance:
          type: integer
        closing_balance:
          type: integer
        received_bags:
          type: integer
        issue_bags:
          type: integer
        reserve_bags:
          type: integer
        chargeable_bags:
          type: integer
      required:
        - commodity_id
    BillVersion:
      type: object
      properties:
        id:
          type: integer
        bill_id:
          type: integer
        data_json:
          type: object
          additionalProperties: true
        version_type:
          type: string
          enum: [edit, delete]
        status:
          type: string
          enum: [pending, approved, rejected]
        created_by:
          type: integer
        created_at:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/User"
        bill:
          oneOf:
            - $ref: "#/components/schemas/Bill"
            - type: "null"
      required:
        - id
        - bill_id
        - data_json
        - version_type
        - status
        - created_by
        - created_at
        - creator
    RequestEditRequest:
      type: object
      properties:
        district:
          type: string
        branch_name:
          type: string
        godown_name:
          type: string
        bill_no:
          type: string
        commodity_id:
          type: integer
        crop_year:
          type: string
        financial_year:
          type: string
        month_year:
          type: string
        rate_per_bag:
          type: number
        opening_balance:
          type: integer
        closing_balance:
          type: integer
        received_bags:
          type: integer
        issue_bags:
          type: integer
        reserve_bags:
          type: integer
        chargeable_bags:
          type: integer
    BillApproval:
      type: object
      properties:
        id:
          type: integer
        bill_id:
          type: integer
        depositor_id:
          type: ["integer", "null"]
        pass_amount:
          type: ["number", "null"]
        payment_method:
          type: ["string", "null"]
        neft_no:
          type: ["string", "null"]
        remark:
          type: ["string", "null"]
        remark_image_url:
          type: ["string", "null"]
        approved_by:
          type: ["integer", "null"]
        approved_at:
          type: ["string", "null"]
          format: date-time
        depositor:
          oneOf:
            - $ref: "#/components/schemas/Depositor"
            - type: "null"
        approver:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
      required:
        - id
        - bill_id
...