# lib/api-zod/src/generated/types/requestEditRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * M.P. Warehousing & Logistics Corporation API
 * OpenAPI spec version: 0.1.0
 */

export interface RequestEditRequest {
  district?: string;
  branch_name?: string;
  godown_name?: string;
  bill_no?: string;
  commodity_id?: number;
  crop_year?: string;
  financial_year?: string;
  month_year?: string;
  rate_per_bag?: number;
  opening_balance?: number;
  closing_balance?: number;
  received_bags?: number;
  issue_bags?: number;
  reserve_bags?: number;
  chargeable_bags?: number;
}